Move PersonContainsRecentPredicate to Model::Person
From Model
